Move common code into SingleValueSensor
This code has been tested in the browser but not on the phone.
