+++ appinventor/components/src/com/google/appinventor/components/runtime/LightSensor.java
@@ -6,15 +6,12 @@
 package com.google.appinventor.components.runtime;
 
 import com.google.appinventor.components.annotations.DesignerComponent;
-import com.google.appinventor.components.annotations.DesignerProperty;
 import com.google.appinventor.components.annotations.PropertyCategory;
 import com.google.appinventor.components.annotations.SimpleEvent;
 import com.google.appinventor.components.annotations.SimpleFunction;
 import com.google.appinventor.components.annotations.SimpleObject;
 import com.google.appinventor.components.annotations.SimpleProperty;
 import com.google.appinventor.components.common.ComponentCategory;
-import com.google.appinventor.components.common.PropertyTypeConstants;
-import com.google.appinventor.components.runtime.util.ErrorMessages;
 
 import android.content.Context;
 import android.hardware.Sensor;
@@ -57,44 +54,6 @@ public class LightSensor extends BufferedSingleValueSensor {
   @SimpleEvent
   public void LightChanged(float lux) {
     EventDispatcher.dispatchEvent(this, "LightChanged", lux);
-  }
-
-  /**
-   * Available property getter method (read-only property).
-   *
-   * @return {@code true} indicates that a light sensor is available,
-   *         {@code false} that it isn't
-   */
-  @SimpleProperty(category = PropertyCategory.BEHAVIOR)
-  public boolean Available() {
-    return isAvailable();
-  }
-
-  /**
-   * If true, the sensor will generate events.  Otherwise, no events
-   * are generated.
-   *
-   * @return {@code true} indicates that the sensor generates events,
-   *         {@code false} that it doesn't
-   */
-  @SimpleProperty(category = PropertyCategory.BEHAVIOR)
-  public boolean Enabled() {
-    return enabled;
-  }
-
-  /**
-   * Specifies whether the sensor should generate events.  If true,
-   * the sensor will generate events.  Otherwise, no events are
-   * generated.
-   *
-   * @param enabled  {@code true} enables sensor event generation,
-   *                 {@code false} disables it
-   */
-  @DesignerProperty(editorType = PropertyTypeConstants.PROPERTY_TYPE_BOOLEAN,
-      defaultValue = "True")
-  @SimpleProperty
-  public void Enabled(boolean enabled) {
-    setEnabled(enabled);
   }
 
   /**
